Use vCloud Tools Tester for launcher integration
Instead of requiring export of environment variables, vCloud Tools Tester reads the variables from a config file. This could be checked into a private repository to allow testing configuration to be shared. It also means that parameters from different tests will not conflict with each if accidentally left over in the environment.