This is called a vApp Template in vCloud Director
For some reason it has been called `catalog_item` in all our configuration files. The plan is to change this to be `vapp_template` as it is in vCloud Director, and there is a story for this: https://www.pivotaltracker.com/story/show/70452106, but changing the test code now instead of waiting for that, as that doesn't affect the actual config.