We do not want to accidentally commit test vars
Instead of providing a default which you then overwrite with the parameters specific to your project, I have instead provided a template file. The idea is that you copy that to `spec/integration` (i.e. next to the template file) and put in your environment details.

For us, this will just be a case of using the file we will have stored in a private repo.

This commit also adds that file to .gitignore so there is no danger of committing your environment's parameters (while doing this I noticed a lot of junk in .gitignore - will leave fixing that for another PR) and updates the tests to use the new location.

There is a corresponding change in vCloud Tools Tester which will warn you if the required file is missing so that it's clear why your test is failing.

I have not added documentation on how to do this - I will add documentation all at once in a separate commit, for clarity.